requirements.txt
black==22.3.0
certifi==2021.10.8
click==8.1.3
elastic-transport==8.1.2
elasticsearch==8.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
